Finished exercise 4.21

+++ part4/blog-list/controllers/bloglist.js
@@ -4,15 +4,6 @@ require('express-async-errors')
 const User = require('../models/userSchema')
 const jwt = require('jsonwebtoken')
 
-const getTokenFrom = (request) => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-
-  return null
-}
-
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
@@ -42,6 +33,17 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if(!decodedToken.id) {
+    return response.status(401).json({error: 'token invalid'})
+  }
+  const blog = await Blog.findById(request.params.id)
+  if(!blog) {
+    return response.status(410).json({error: 'blog already deleted'})
+  }
+  if(blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({error: 'can only delete own blogs'})
+  }
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
